Module stm32l4x6::adc123_common

Analog-to-Digital Converter

Modules

ccr
ADC common control register

cdr
ADC common regular data register for dual and triple modes

csr
ADC Common status register

Structs

CCR
ADC common control register

CDR
ADC common regular data register for dual and triple modes

CSR
ADC Common status register

RegisterBlock
Register block