Struct stm32l4x6::adc123_common::csr::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn addrdy_mst(&self) -> ADDRDY_MSTR

Bit 0 - ADDRDY_MST

pub fn eosmp_mst(&self) -> EOSMP_MSTR

Bit 1 - EOSMP_MST

pub fn eoc_mst(&self) -> EOC_MSTR

Bit 2 - EOC_MST

pub fn eos_mst(&self) -> EOS_MSTR

Bit 3 - EOS_MST

pub fn ovr_mst(&self) -> OVR_MSTR

Bit 4 - OVR_MST

pub fn jeoc_mst(&self) -> JEOC_MSTR

Bit 5 - JEOC_MST

pub fn jeos_mst(&self) -> JEOS_MSTR

Bit 6 - JEOS_MST

pub fn awd1_mst(&self) -> AWD1_MSTR

Bit 7 - AWD1_MST

pub fn awd2_mst(&self) -> AWD2_MSTR

Bit 8 - AWD2_MST

pub fn awd3_mst(&self) -> AWD3_MSTR

Bit 9 - AWD3_MST

pub fn jqovf_mst(&self) -> JQOVF_MSTR

Bit 10 - JQOVF_MST

pub fn adrdy_slv(&self) -> ADRDY_SLVR

Bit 16 - ADRDY_SLV

pub fn eosmp_slv(&self) -> EOSMP_SLVR

Bit 17 - EOSMP_SLV

pub fn eoc_slv(&self) -> EOC_SLVR

Bit 18 - End of regular conversion of the slave ADC

pub fn eos_slv(&self) -> EOS_SLVR

Bit 19 - End of regular sequence flag of the slave ADC

pub fn ovr_slv(&self) -> OVR_SLVR

Bit 20 - Overrun flag of the slave ADC

pub fn jeoc_slv(&self) -> JEOC_SLVR

Bit 21 - End of injected conversion flag of the slave ADC

pub fn jeos_slv(&self) -> JEOS_SLVR

Bit 22 - End of injected sequence flag of the slave ADC

pub fn awd1_slv(&self) -> AWD1_SLVR

Bit 23 - Analog watchdog 1 flag of the slave ADC

pub fn awd2_slv(&self) -> AWD2_SLVR

Bit 24 - Analog watchdog 2 flag of the slave ADC

pub fn awd3_slv(&self) -> AWD3_SLVR

Bit 25 - Analog watchdog 3 flag of the slave ADC

pub fn jqovf_slv(&self) -> JQOVF_SLVR

Bit 26 - Injected Context Queue Overflow flag of the slave ADC