Struct stm32l4x6::gpiob::afrl::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn afrl7(&mut self) -> _AFRL7W

Bits 28:31 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl6(&mut self) -> _AFRL6W

Bits 24:27 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl5(&mut self) -> _AFRL5W

Bits 20:23 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl4(&mut self) -> _AFRL4W

Bits 16:19 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl3(&mut self) -> _AFRL3W

Bits 12:15 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl2(&mut self) -> _AFRL2W

Bits 8:11 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl1(&mut self) -> _AFRL1W

Bits 4:7 - Alternate function selection for port x bit y (y = 0..7)

pub fn afrl0(&mut self) -> _AFRL0W

Bits 0:3 - Alternate function selection for port x bit y (y = 0..7)