Struct stm32l4x6::rcc::ahb2smenr::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn rngsmen(&self) -> RNGSMENR

Bit 18 - Random Number Generator clocks enable during Sleep and Stop modes

pub fn aessmen(&self) -> AESSMENR

Bit 16 - AES accelerator clocks enable during Sleep and Stop modes

pub fn adcfssmen(&self) -> ADCFSSMENR

Bit 13 - ADC clocks enable during Sleep and Stop modes

pub fn otgfssmen(&self) -> OTGFSSMENR

Bit 12 - OTG full speed clocks enable during Sleep and Stop modes

pub fn sram2smen(&self) -> SRAM2SMENR

Bit 9 - SRAM2 interface clocks enable during Sleep and Stop modes

pub fn gpiohsmen(&self) -> GPIOHSMENR

Bit 7 - IO port H clocks enable during Sleep and Stop modes

pub fn gpiogsmen(&self) -> GPIOGSMENR

Bit 6 - IO port G clocks enable during Sleep and Stop modes

pub fn gpiofsmen(&self) -> GPIOFSMENR

Bit 5 - IO port F clocks enable during Sleep and Stop modes

pub fn gpioesmen(&self) -> GPIOESMENR

Bit 4 - IO port E clocks enable during Sleep and Stop modes

pub fn gpiodsmen(&self) -> GPIODSMENR

Bit 3 - IO port D clocks enable during Sleep and Stop modes

pub fn gpiocsmen(&self) -> GPIOCSMENR

Bit 2 - IO port C clocks enable during Sleep and Stop modes

pub fn gpiobsmen(&self) -> GPIOBSMENR

Bit 1 - IO port B clocks enable during Sleep and Stop modes

pub fn gpioasmen(&self) -> GPIOASMENR

Bit 0 - IO port A clocks enable during Sleep and Stop modes