Struct stm32l4x6::rcc::apb1smenr1::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn lptim1smen(&self) -> LPTIM1SMENR

Bit 31 - Low power timer 1 clocks enable during Sleep and Stop modes

pub fn opampsmen(&self) -> OPAMPSMENR

Bit 30 - OPAMP interface clocks enable during Sleep and Stop modes

pub fn dac1smen(&self) -> DAC1SMENR

Bit 29 - DAC1 interface clocks enable during Sleep and Stop modes

pub fn pwrsmen(&self) -> PWRSMENR

Bit 28 - Power interface clocks enable during Sleep and Stop modes

pub fn can1smen(&self) -> CAN1SMENR

Bit 25 - CAN1 clocks enable during Sleep and Stop modes

pub fn i2c3smen(&self) -> I2C3SMENR

Bit 23 - I2C3 clocks enable during Sleep and Stop modes

pub fn i2c2smen(&self) -> I2C2SMENR

Bit 22 - I2C2 clocks enable during Sleep and Stop modes

pub fn i2c1smen(&self) -> I2C1SMENR

Bit 21 - I2C1 clocks enable during Sleep and Stop modes

pub fn uart5smen(&self) -> UART5SMENR

Bit 20 - UART5 clocks enable during Sleep and Stop modes

pub fn uart4smen(&self) -> UART4SMENR

Bit 19 - UART4 clocks enable during Sleep and Stop modes

pub fn usart3smen(&self) -> USART3SMENR

Bit 18 - USART3 clocks enable during Sleep and Stop modes

pub fn usart2smen(&self) -> USART2SMENR

Bit 17 - USART2 clocks enable during Sleep and Stop modes

pub fn sp3smen(&self) -> SP3SMENR

Bit 15 - SPI3 clocks enable during Sleep and Stop modes

pub fn spi2smen(&self) -> SPI2SMENR

Bit 14 - SPI2 clocks enable during Sleep and Stop modes

pub fn wwdgsmen(&self) -> WWDGSMENR

Bit 11 - Window watchdog clocks enable during Sleep and Stop modes

pub fn lcdsmen(&self) -> LCDSMENR

Bit 9 - LCD clocks enable during Sleep and Stop modes

pub fn tim7smen(&self) -> TIM7SMENR

Bit 5 - TIM7 timer clocks enable during Sleep and Stop modes

pub fn tim6smen(&self) -> TIM6SMENR

Bit 4 - TIM6 timer clocks enable during Sleep and Stop modes

pub fn tim5smen(&self) -> TIM5SMENR

Bit 3 - Reserved

pub fn tim4smen(&self) -> TIM4SMENR

Bit 2 - TIM4 timer clocks enable during Sleep and Stop modes

pub fn tim3smen(&self) -> TIM3SMENR

Bit 1 - TIM3 timer clocks enable during Sleep and Stop modes

pub fn tim2smen(&self) -> TIM2SMENR

Bit 0 - TIM2 timer clocks enable during Sleep and Stop modes