Struct stm32l4x6::rcc::ahb2smenr::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn rngsmen(&mut self) -> _RNGSMENW

Bit 18 - Random Number Generator clocks enable during Sleep and Stop modes

pub fn aessmen(&mut self) -> _AESSMENW

Bit 16 - AES accelerator clocks enable during Sleep and Stop modes

pub fn adcfssmen(&mut self) -> _ADCFSSMENW

Bit 13 - ADC clocks enable during Sleep and Stop modes

pub fn otgfssmen(&mut self) -> _OTGFSSMENW

Bit 12 - OTG full speed clocks enable during Sleep and Stop modes

pub fn sram2smen(&mut self) -> _SRAM2SMENW

Bit 9 - SRAM2 interface clocks enable during Sleep and Stop modes

pub fn gpiohsmen(&mut self) -> _GPIOHSMENW

Bit 7 - IO port H clocks enable during Sleep and Stop modes

pub fn gpiogsmen(&mut self) -> _GPIOGSMENW

Bit 6 - IO port G clocks enable during Sleep and Stop modes

pub fn gpiofsmen(&mut self) -> _GPIOFSMENW

Bit 5 - IO port F clocks enable during Sleep and Stop modes

pub fn gpioesmen(&mut self) -> _GPIOESMENW

Bit 4 - IO port E clocks enable during Sleep and Stop modes

pub fn gpiodsmen(&mut self) -> _GPIODSMENW

Bit 3 - IO port D clocks enable during Sleep and Stop modes

pub fn gpiocsmen(&mut self) -> _GPIOCSMENW

Bit 2 - IO port C clocks enable during Sleep and Stop modes

pub fn gpiobsmen(&mut self) -> _GPIOBSMENW

Bit 1 - IO port B clocks enable during Sleep and Stop modes

pub fn gpioasmen(&mut self) -> _GPIOASMENW

Bit 0 - IO port A clocks enable during Sleep and Stop modes