Struct stm32l4x6::rcc::apb1smenr1::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn lptim1smen(&mut self) -> _LPTIM1SMENW

Bit 31 - Low power timer 1 clocks enable during Sleep and Stop modes

pub fn opampsmen(&mut self) -> _OPAMPSMENW

Bit 30 - OPAMP interface clocks enable during Sleep and Stop modes

pub fn dac1smen(&mut self) -> _DAC1SMENW

Bit 29 - DAC1 interface clocks enable during Sleep and Stop modes

pub fn pwrsmen(&mut self) -> _PWRSMENW

Bit 28 - Power interface clocks enable during Sleep and Stop modes

pub fn can1smen(&mut self) -> _CAN1SMENW

Bit 25 - CAN1 clocks enable during Sleep and Stop modes

pub fn i2c3smen(&mut self) -> _I2C3SMENW

Bit 23 - I2C3 clocks enable during Sleep and Stop modes

pub fn i2c2smen(&mut self) -> _I2C2SMENW

Bit 22 - I2C2 clocks enable during Sleep and Stop modes

pub fn i2c1smen(&mut self) -> _I2C1SMENW

Bit 21 - I2C1 clocks enable during Sleep and Stop modes

pub fn uart5smen(&mut self) -> _UART5SMENW

Bit 20 - UART5 clocks enable during Sleep and Stop modes

pub fn uart4smen(&mut self) -> _UART4SMENW

Bit 19 - UART4 clocks enable during Sleep and Stop modes

pub fn usart3smen(&mut self) -> _USART3SMENW

Bit 18 - USART3 clocks enable during Sleep and Stop modes

pub fn usart2smen(&mut self) -> _USART2SMENW

Bit 17 - USART2 clocks enable during Sleep and Stop modes

pub fn sp3smen(&mut self) -> _SP3SMENW

Bit 15 - SPI3 clocks enable during Sleep and Stop modes

pub fn spi2smen(&mut self) -> _SPI2SMENW

Bit 14 - SPI2 clocks enable during Sleep and Stop modes

pub fn wwdgsmen(&mut self) -> _WWDGSMENW

Bit 11 - Window watchdog clocks enable during Sleep and Stop modes

pub fn lcdsmen(&mut self) -> _LCDSMENW

Bit 9 - LCD clocks enable during Sleep and Stop modes

pub fn tim7smen(&mut self) -> _TIM7SMENW

Bit 5 - TIM7 timer clocks enable during Sleep and Stop modes

pub fn tim6smen(&mut self) -> _TIM6SMENW

Bit 4 - TIM6 timer clocks enable during Sleep and Stop modes

pub fn tim5smen(&mut self) -> _TIM5SMENW

Bit 3 - Reserved

pub fn tim4smen(&mut self) -> _TIM4SMENW

Bit 2 - TIM4 timer clocks enable during Sleep and Stop modes

pub fn tim3smen(&mut self) -> _TIM3SMENW

Bit 1 - TIM3 timer clocks enable during Sleep and Stop modes

pub fn tim2smen(&mut self) -> _TIM2SMENW

Bit 0 - TIM2 timer clocks enable during Sleep and Stop modes